Inverse order of the UI
Inverse order of the UI,
so that the most important barcode image and result comes first.
Gallery comes before custom image uploading,
as most people are likely to click through the gallery at most.

Also, scrolling to the top is less surprising as people have already
seen the result we are scrolling to.

Fix height and reserve place for the barcode image and messages,
so that the page doesn't jump due to varying height of messages and images.